Support for landscape layout. Support for configuration changes.

+++ app/src/main/java/com/lanyard/pirateerlite/models/JobModel.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2019 Peter Respondek
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.lanyard.pirateerlite.models
 
 import com.lanyard.library.Graph
@@ -8,17 +24,17 @@ import com.lanyard.pirateerlite.singletons.Map
 class JobModel {
     var source: TownModel
 
-        get() { return Map.instance.towns[_data.sourceTownId] }
+        get() { return Map.instance.towns[_data.sourceTownId.toInt()] }
         set(town) {
-            _data.destinationTownId = Map.instance.towns.indexOfFirst { it === town }
+            _data.destinationTownId = Map.instance.towns.indexOfFirst { it === town }.toLong()
             calcValue()
         }
 
     var destination: TownModel
 
-        get() { return Map.instance.towns[_data.destinationTownId] }
+        get() { return Map.instance.towns[_data.destinationTownId.toInt()] }
         set(town) {
-            _data.destinationTownId = Map.instance.towns.indexOfFirst { it === town }
+            _data.destinationTownId = Map.instance.towns.indexOfFirst { it === town }.toLong()
             calcValue()
         }
 
@@ -54,8 +70,8 @@ class JobModel {
             JobController.jobData.random()[0],
             0.0f,
             Math.random() * 10.0 <= 1.0,
-            Map.instance.towns.indexOfFirst { it === source },
-            Map.instance.towns.indexOfFirst { it === destination },
+            Map.instance.towns.indexOfFirst { it === source }.toLong(),
+            Map.instance.towns.indexOfFirst { it === destination }.toLong(),
             1.0
         )
         calcValue()
